Compute and store extended sequence numbers as Long rather than Int.

+++ jitsi-media-transform/src/main/kotlin/org/jitsi/nlj/rtp/TransportCcEngine.kt
@@ -329,7 +329,7 @@ class TransportCcEngine(
         }
 
         val lastSequence: Int
-            get() = if (lastIndex == -1) -1 else lastIndex and 0xFFFF
+            get() = if (lastIndex == -1L) -1 else (lastIndex and 0xFFFF).toInt()
     }
 
     companion object {
